changed dir for image creation

+++ main.py
@@ -50,24 +50,28 @@ if connection_code == 200:
     data = ast.literal_eval(create_dataset(args.json_file).stdout.decode("UTF-8"))
     print('Status: ' + data['status'])
 
-    print('Preparing files...')
-    zip_files_list = []
+    if data['status'] == 'OK':
+        print('Preparing files...')
+        zip_files_list = []
 
-    zip_files_list.append(create_image(args.dir, args.files_prefix))
-    
-    zip_files_list.extend(zip_files(args.dir, args.files_prefix, args.files_prefix))
-    print('Done')
+        zip_files_list.append(create_image(args.dir, args.files_prefix))
+        
+        zip_files_list.extend(zip_files(args.dir, args.files_prefix, args.files_prefix))
+        print('Done')
 
-    print('Sending zips to dataset...')
-    for it, file in enumerate(zip_files_list):
-        print('Uploading file #' + str(it))
-        code = upload_file_to_dataset(data['data']['persistentId'], file, ZIP_FILES_DESCRIPTION)
-        print('Status: ' + str(code))
-    print('Done')
+        print('Sending zips to dataset...')
+        for it, file in enumerate(zip_files_list):
+            print('Uploading file #' + str(it))
+            code = upload_file_to_dataset(data['data']['persistentId'], file, ZIP_FILES_DESCRIPTION)
+            print('Status: ' + str(code))
+        print('Done')
 
-    print('Cleaning...')
-    for file in zip_files_list:
-        os.remove(file)
-    print('Done')
+        print('Cleaning...')
+        for file in zip_files_list:
+            os.remove(file)
+        print('Done')
+    else:
+        print('There was a problem with dataset creation:')
+        print(data)
 else:
     print('Script could not connect to dataverse!')
